refactor users, user tests, permission mixins; add statuses with tests

+++ task_manager/users/tests.py
@@ -2,9 +2,10 @@ from django.urls import reverse_lazy
 from django.test import TestCase, Client
 from .models import CustomUser
 from .fixtures.fixtures import valid_users, invalid_users, extended_valid_users
+from task_manager._test_utils._test_utils import _TestUtilsMixin
 
 
-class _TestUtilsMixin:
+class _TestUsersUtilsMixin(_TestUtilsMixin):
     def delete_user(self, username):
         if username:
             if user := CustomUser.objects.filter(username=username):
@@ -17,8 +18,9 @@ class _TestUtilsMixin:
             )
         [self.delete_user(user["username"]) for user in users]
 
-    def create_user(self, user_data):
-        self.delete_user(user_data["username"])
+    def create_user(self, user_data, force=False):
+        if force:
+            self.delete_user(user_data["username"])
         user = CustomUser.objects.get_or_create(
             username=user_data["username"],
             password=user_data.get("password")
@@ -29,35 +31,18 @@ class _TestUtilsMixin:
         )
         return user
 
-    def _test_users(self, test_data):
-        for test_name, data in test_data:
-            test = getattr(self, test_name)
-            [test(user) for user in data]
-
-    def _test_users_to_success(self):
-        print("\n\nStarting tests for users post views to succeed.")
-        self._test_users(self.tests_to_success)
-
-    def _test_users_to_fail(self):
-        print("\n\nStarting tests for users post views to fail.")
-        self._test_users(self.tests_to_fail)
-
-    def tests_all(self):
-        # self._test_users_to_success()
-        # self._test_users_to_fail()
-        pass
-
 
 # Create your tests here.
-class TestUsersPostCRUDSs(TestCase, _TestUtilsMixin):
+class TestUsersPostCRUDSs(TestCase, _TestUsersUtilsMixin):
     def setUp(self):
+        self.subject = "Users app (post method)"
         self.client = Client()
-        self.tests_to_success = (
+        self._tests_to_success_tuple = (
             ("_test_users_create_post_success", valid_users),
             ("_test_users_update_post_success", valid_users),
             ("_test_users_delete_post_success", valid_users),
         )
-        self.tests_to_fail = (
+        self._tests_to_fail_tuple = (
             ("_test_users_create_post_fail", invalid_users),
             ("_test_users_create_by_user_post_fail", valid_users),
             ("_test_users_create_existing_user_post_fail", valid_users),
@@ -82,6 +67,7 @@ class TestUsersPostCRUDSs(TestCase, _TestUtilsMixin):
         )
         response = self.client.post(reverse_lazy("users_create"), data)
         self.assertEqual(response.status_code, 302)
+        self.assertRedirects(response, reverse_lazy("index"))
         self.assertTrue(
             CustomUser.objects.filter(username=data["username"]).exists()
         )
@@ -100,6 +86,7 @@ class TestUsersPostCRUDSs(TestCase, _TestUtilsMixin):
         url = reverse_lazy("users_update", args=(user["id"],))
         response = self.client.post(url, data)
         self.assertEqual(response.status_code, 302)
+        self.assertRedirects(response, reverse_lazy("users_index"))
         updated_user = CustomUser.objects.get(username=data["username"])
         self.assertEqual(
             (updated_user.first_name, updated_user.last_name),
@@ -114,6 +101,7 @@ class TestUsersPostCRUDSs(TestCase, _TestUtilsMixin):
         url = reverse_lazy("users_delete", args=(id,))
         response = self.client.post(url)
         self.assertEqual(response.status_code, 302)
+        self.assertRedirects(response, reverse_lazy("users_index"))
         self.assertFalse(
             CustomUser.objects.filter(username=user["username"]).exists()
         )
@@ -144,7 +132,7 @@ class TestUsersPostCRUDSs(TestCase, _TestUtilsMixin):
             "new_password1": user["password1"],
             "new_password2": user["password2"],
         }
-        self.create_user(data)
+        self.create_user(data, force=True)
         print(
             "Test create existing user (POST) with data to fail: ",
             data,
@@ -161,14 +149,8 @@ class TestUsersPostCRUDSs(TestCase, _TestUtilsMixin):
         print("successfully failed.")
 
     def _test_users_create_by_user_post_fail(self, user):
-        data = {
-            "username": user["username"] + "x",
-            "first_name": user["last_name"],
-            "last_name": user["first_name"],
-            "new_password1": user["password1"],
-            "new_password2": user["password2"],
-        }
-        self.create_user(user)
+        data = user | {"username": user["username"] + "x"}
+        self.create_user(user, force=True)
         print(
             "Test create existing user by user (POST) with data to fail: ",
             data,
@@ -182,15 +164,16 @@ class TestUsersPostCRUDSs(TestCase, _TestUtilsMixin):
         )
         self.client.login(username=user["username"], password=user["password1"])
         response = self.client.post(reverse_lazy("users_create"), data)
-        self.assertEqual(response.status_code, 200)
-        self.assertFalse(
-            CustomUser.objects.filter(username=data["username"]).exists()
+        self.assertEqual(response.status_code, 302)
+        self.assertRedirects(response, reverse_lazy("index"))
+        self.assertTrue(
+            CustomUser.objects.filter(username=user["username"]).exists()
         )
         print("successfully failed.")
 
     def _test_users_update_post_fail(self, data):
         valid_user_data = data[0]
-        valid_user, _ = self.create_user(valid_user_data)
+        valid_user, _ = self.create_user(valid_user_data, force=True)
         invalid_user_data = data[1]
         test_data = {
             "username": invalid_user_data["username"],
@@ -202,9 +185,11 @@ class TestUsersPostCRUDSs(TestCase, _TestUtilsMixin):
         print("Test update user with broken data: ", test_data, end=" => ")
 
         print("Test updating data without authentication")
+        self.client.logout()
         url = reverse_lazy("users_update", args=(valid_user.id,))
         response = self.client.post(url, test_data)
         self.assertEqual(response.status_code, 302)
+        self.assertRedirects(response, reverse_lazy("users_login"))
         updated_user = CustomUser.objects.get(username=valid_user.username)
         self.assertEqual(
             (updated_user.first_name, updated_user.last_name),
@@ -218,6 +203,7 @@ class TestUsersPostCRUDSs(TestCase, _TestUtilsMixin):
         url = reverse_lazy("users_update", args=(valid_user.id,))
         response = self.client.post(url, test_data)
         self.assertEqual(response.status_code, 302)
+        self.assertRedirects(response, reverse_lazy("users_update"))
         updated_user = CustomUser.objects.get(username=valid_user.username)
         self.assertEqual(
             (updated_user.first_name, updated_user.last_name),
@@ -228,10 +214,11 @@ class TestUsersPostCRUDSs(TestCase, _TestUtilsMixin):
         self.client.login(
             username=valid_user.username, password=valid_user.password
         )
-        invalid_id = list({1, 2, 3} ^ {valid_user.id})[0]
-        url = reverse_lazy("users_update", args=(invalid_id,))
+        other_users_id = list({1, 2, 3} ^ {valid_user.id})[0]
+        url = reverse_lazy("users_update", args=(other_users_id,))
         response = self.client.post(url, test_data)
         self.assertEqual(response.status_code, 302)
+        self.assertRedirects(response, reverse_lazy("users_index"))
         updated_user = CustomUser.objects.get(username=valid_user.username)
         self.assertEqual(
             (updated_user.first_name, updated_user.last_name),
@@ -240,7 +227,7 @@ class TestUsersPostCRUDSs(TestCase, _TestUtilsMixin):
         print("All successfully failed.")
 
     def _test_users_delete_post_fail(self, user_data):
-        valid_user, _ = self.create_user(user_data)
+        valid_user, _ = self.create_user(user_data, force=True)
 
         print("Test delete user: ", user_data, end=" => ")
 
@@ -248,19 +235,7 @@ class TestUsersPostCRUDSs(TestCase, _TestUtilsMixin):
         url = reverse_lazy("users_delete", args=(valid_user.id,))
         response = self.client.post(url)
         self.assertEqual(response.status_code, 302)
-        updated_user = CustomUser.objects.get(username=valid_user.username)
-        self.assertEqual(
-            (updated_user.first_name, updated_user.last_name),
-            (valid_user.first_name, valid_user.last_name),
-        )
-
-        print("Test deleting data with bad data")
-        self.client.login(
-            username=valid_user.username, password=valid_user.password
-        )
-        url = reverse_lazy("users_delete", args=(valid_user.id,))
-        response = self.client.post(url)
-        self.assertEqual(response.status_code, 302)
+        self.assertRedirects(response, reverse_lazy("users_login"))
         updated_user = CustomUser.objects.get(username=valid_user.username)
         self.assertEqual(
             (updated_user.first_name, updated_user.last_name),
@@ -275,6 +250,7 @@ class TestUsersPostCRUDSs(TestCase, _TestUtilsMixin):
         url = reverse_lazy("users_delete", args=(invalid_id,))
         response = self.client.post(url)
         self.assertEqual(response.status_code, 302)
+        self.assertRedirects(response, reverse_lazy("users_index"))
         updated_user = CustomUser.objects.get(username=valid_user.username)
         self.assertEqual(
             (updated_user.first_name, updated_user.last_name),
@@ -283,29 +259,30 @@ class TestUsersPostCRUDSs(TestCase, _TestUtilsMixin):
         print("All successfully failed.")
 
 
-class TestUsersGetCRUDSs(TestCase, _TestUtilsMixin):
+class TestUsersGetCRUDSs(TestCase, _TestUsersUtilsMixin):
     def setUp(self):
+        self.subject = "Users app (get method)"
         self.client = Client()
-        self.tests_to_success = (
-            ("_test_users_create_get_success", valid_users),
-            #    ("_test_users_update_get_success", valid_users),
-            #    ("_test_users_delete_get_success", valid_users),
+        self._tests_to_success_tuple = (
+            ("_test_users_index_get_success", valid_users),
         )
-        self.tests_to_fail = (
-            ("_test_users_create_get_fail", invalid_users),
-            (
-                "_test_users_update_get_fail",
-                tuple(zip(valid_users * 3, invalid_users)),
-            ),
-            ("_test_users_delete_get_fail", valid_users),
-        )
+        self._tests_to_fail_tuple = []
 
-    def tests_all(self):
-        pass
-
-    def _test_users_create_get_success(self):
-        print("Test user registration page (GET)", end=" => ")
-        response = self.client.get(reverse_lazy("users_create"))
+    def _test_users_index_get_success(self, user):
+        response = self.client.get(reverse_lazy("users_index"))
         self.assertEqual(response.status_code, 200)
-        self.assertContains(response)
+        self.assertNotContains(response, user["username"])
+        self.assertNotContains(response, user["first_name"])
+        self.assertNotContains(response, user["last_name"])
+        self.create_user(user, force=True)
+        print(
+            "Test user registration page (GET) with user data ",
+            user,
+            end=" => ",
+        )
+        response = self.client.get(reverse_lazy("users_index"))
+        self.assertEqual(response.status_code, 200)
+        self.assertContains(response, user["username"])
+        self.assertContains(response, user["first_name"])
+        self.assertContains(response, user["last_name"])
         print("succeeded.")
